examples/story_generator/requirements.txt
guidance
python-dotenv